weeks/week-17/events-s18/gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
strawberry-graphql>=0.247.0

